refactor: extract out common code to `AbstractIsochroneMapBuilder`
